backend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
certifi
google-generativeai
pypdf
